+++ src/lib/config.ts
@@ -62,8 +62,8 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "1.0.0",
-  sdkVersion: "0.0.1-alpha.4",
-  genVersion: "2.656.3",
+  sdkVersion: "0.0.1-alpha.5",
+  genVersion: "2.657.1",
   userAgent:
-    "speakeasy-sdk/typescript 0.0.1-alpha.4 2.656.3 1.0.0 mollie-api-typescript",
+    "speakeasy-sdk/typescript 0.0.1-alpha.5 2.657.1 1.0.0 mollie-api-typescript",
 } as const;